Move updateCourse action to own file
